# frontend/src/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "RouteRank API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/api/routes/{id}" : {
      "put" : {
        "tags" : [ "routes-controller" ],
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RouteInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RouteView"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "routes-controller" ],
        "operationId" : "remove",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/admin/bus-routes/{id}" : {
      "put" : {
        "tags" : [ "bus-routes-admin-controller" ],
        "operationId" : "redraw",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BusRouteInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/admin/bus-routes" : {
      "get" : {
        "tags" : [ "bus-routes-admin-controller" ],
        "operationId" : "list_1",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/BusRouteView"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "bus-routes-admin-controller" ],
        "operationId" : "add",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BusRouteInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Created"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/bus-routes/{id}/retire" : {
      "post" : {
        "tags" : [ "bus-routes-admin-controller" ],
        "operationId" : "retire",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Reason"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
    "/api/admin/bus-routes/draw" : {
      "post" : {
        "tags" : [ "bus-routes-admin-controller" ],
        "operationId" : "draw",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BusRouteInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DrawingView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "LatLon" : {
        "type" : "object",
        "properties" : {
          "lat" : {
            "type" : "number",
            "format" : "double",
            "maximum" : 10.2,
            "minimum" : 5.5
          },
          "lon" : {
            "type" : "number",
            "format" : "double",
            "maximum" : 82.2,
            "minimum" : 79.3
          }
        }
      },
      "RouteInput" : {
        "type" : "object",
        "properties" : {
          "start" : {
            "$ref" : "#/components/schemas/LatLon"
          },
          "end" : {
            "$ref" : "#/components/schemas/LatLon"
          },
          "waypoints" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LatLon"
            },
            "maxItems" : 8,
            "minItems" : 0
          },
          "slot" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 3,
            "minimum" : 1
          },
          "routeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "extend" : {
            "type" : "integer",
            "format" : "int64"
          },
          "turnstile" : {
            "type" : "string",
            "maxLength" : 2048,
            "minLength" : 0
          },
          "device" : {
            "type" : "string",
            "maxLength" : 64,
            "minLength" : 0
          },
          "website" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          }
        },
        "required" : [ "end", "start", "waypoints" ]
      },
      "BusRef" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "number" : {
            "type" : "string"
          }
        }
      },
      "Busiest" : {
        "type" : "object",
        "properties" : {
          "slug" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "points" : {
            "type" : "integer",
            "format" : "int32"
          },
          "people" : {
            "type" : "integer",
            "format" : "int32"
          },
          "rank" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "RouteView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "slot" : {
            "type" : "integer",
            "format" : "int32"
          },
          "name" : {
            "type" : "string"
          },
          "start" : {
            "$ref" : "#/components/schemas/LatLon"
          },
          "end" : {
            "$ref" : "#/components/schemas/LatLon"
          },
          "waypoints" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LatLon"
            }
          },
          "out" : {
            "type" : "array",
            "items" : {
              "type" : "array",
              "items" : {
                "type" : "number",
                "format" : "double"
              }
            }
          },
          "back" : {
            "type" : "array",
            "items" : {
              "type" : "array",
              "items" : {
                "type" : "number",
                "format" : "double"
              }
            }
          },
          "lengthOutM" : {
            "type" : "number",
            "format" : "double"
          },
          "lengthBackM" : {
            "type" : "number",
            "format" : "double"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "busiest" : {
            "$ref" : "#/components/schemas/Busiest"
          },
          "extendsBus" : {
            "$ref" : "#/components/schemas/BusRef"
          }
        },
        "required" : [ "back", "createdAt", "end", "id", "lengthBackM", "lengthOutM", "name", "out", "slot", "start", "updatedAt", "waypoints" ]
      },
...
      "BusRouteInput" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "number" : {
            "type" : "string",
            "maxLength" : 12,
            "minLength" : 0,
            "pattern" : "[0-9A-Za-z/ -]*"
          },
          "startName" : {
            "type" : "string",
            "maxLength" : 60,
            "minLength" : 0
          },
          "endName" : {
            "type" : "string",
            "maxLength" : 60,
            "minLength" : 0
          },
          "start" : {
            "$ref" : "#/components/schemas/LatLon"
          },
          "end" : {
            "$ref" : "#/components/schemas/LatLon"
          },
          "waypoints" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LatLon"
            },
            "maxItems" : 25,
            "minItems" : 0
          },
          "reason" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          }
        },
        "required" : [ "end", "start", "waypoints" ]
      },
...
      "Reason" : {
        "type" : "object",
        "properties" : {
          "reason" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          }
        },
        "required" : [ "reason" ]
      },
...
      "Created" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "DrawingView" : {
        "type" : "object",
        "properties" : {
          "points" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LatLon"
            }
          },
          "out" : {
            "type" : "array",
            "items" : {
              "type" : "array",
              "items" : {
                "type" : "number",
                "format" : "double"
              }
            }
          },
          "back" : {
            "type" : "array",
            "items" : {
              "type" : "array",
              "items" : {
                "type" : "number",
                "format" : "double"
              }
            }
          },
          "lengthOutM" : {
            "type" : "number",
            "format" : "double"
          },
          "lengthBackM" : {
            "type" : "number",
            "format" : "double"
          },
          "startName" : {
            "type" : "string"
          },
          "endName" : {
            "type" : "string"
          },
          "towns" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "problems" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "BusRouteView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "number" : {
            "type" : "string"
          },
          "startName" : {
            "type" : "string"
          },
          "endName" : {
            "type" : "string"
          },
          "start" : {
            "$ref" : "#/components/schemas/LatLon"
          },
          "end" : {
            "$ref" : "#/components/schemas/LatLon"
          },
          "waypoints" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LatLon"
            }
          },
          "out" : {
            "type" : "array",
            "items" : {
              "type" : "array",
              "items" : {
                "type" : "number",
                "format" : "double"
              }
            }
          },
          "back" : {
            "type" : "array",
            "items" : {
              "type" : "array",
              "items" : {
                "type" : "number",
                "format" : "double"
              }
            }
          },
          "lengthOutM" : {
            "type" : "number",
            "format" : "double"
          },
          "lengthBackM" : {
            "type" : "number",
            "format" : "double"
          },
          "towns" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "retiredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "extensions" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    }
  }
}